+++ __tests__/__unit__/middleware.test.js
@@ -1,10 +1,11 @@
 const { spalatumMiddleware } = require('../../lib');
+const responseMock = require('../../__mocks__/response');
 
 const mockNext = jest.fn();
 
 describe('# When calling express middleware', () => {
   beforeEach(() => {
-    spalatumMiddleware({}, {}, mockNext);
+    spalatumMiddleware({}, responseMock(), mockNext);
   });
 
   it(`Should call the "next" callback function,